[mongodb] separation between MongoCommandSubscriber and Patcher
